requirements.txt
colour-science==0.4.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
imageio==2.22.1
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.9.2
six==1.16.0
typing_extensions==4.4.0
